% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isisinsight-package.R
\docType{package}
\name{isisinsight-package}
\alias{isisinsight}
\alias{isisinsight-package}
\title{isisinsight: ISIS-Fish insight R package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package aims to provide processes for exploring outputs from the ISIS-Fish (Integration of Spatial Information and Simulation for Fisheries) simulation tool. You can find various types of methods, go for example from visualisation or data manipulation, uncertainty analysis, scenarios display or global exploration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umr-marbec/isisinsight}
  \item \url{https://umr-marbec.github.io/isisinsight/}
  \item Report bugs at \url{https://github.com/umr-marbec/isisinsight/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@cnrs.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})

Authors:
\itemize{
  \item Stephanie Mahevas \email{stephanie.mahevas@ifremer.fr}
}

Other contributors:
\itemize{
  \item UMR MARBEC MARine Biodiversity Exploitation & Conservation [copyright holder]
}

}
\keyword{internal}
