% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_simulations_settings.R
\name{outputs_simulations_settings}
\alias{outputs_simulations_settings}
\title{ISIS-Fish outputs simulations formatting}
\usage{
outputs_simulations_settings(directory_path)
}
\arguments{
\item{directory_path}{Mandatory. Class character expected. Directory path of the ISIS-Fish outputs simulations.}
}
\value{
The function returns a list with a length in relation to the number of simulation directory provided. Each element of the list has information about metadata and data (original and improved) associated with the simulation.
}
\description{
Function for ISIS-Fish outputs simulations formatting and manipulation.
}
\examples{
#replace the value of directory_path by a correct path
try(outputs_simulations_settings(directory_path = "my/path/to/simulations/directory"))

}
